add SYMBOL switch for cmake

+++ contrib/agilent/libs/ap240controls/waveform.hpp
@@ -114,7 +114,7 @@ namespace ap240controls {
         template<typename T> const T* end() const;
 
         template<typename T> void advance( const T*& it, size_t distance ) const {
-            it = ( distance && distance < std::distance( it, end<T>() ) ) ? it + distance : end<T>();
+            it = ( distance && distance < size_t( std::distance( it, end<T>() ) ) ? it + distance : end<T>() );
         }
 
         std::pair<double, int> operator [] ( size_t ) const;
